SW-152 Merge `release` into `main`

+++ src/main/java/com/sweep/jaksim31/adapter/cache/MemberCacheAdapter.java
@@ -1,15 +1,12 @@
 package com.sweep.jaksim31.adapter.cache;
 
-import com.google.common.collect.Lists;
-import com.sweep.jaksim31.adapter.RestPage;
-import com.sweep.jaksim31.dto.diary.DiaryInfoResponse;
 import com.sweep.jaksim31.dto.member.MemberInfoResponse;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.*;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.core.ValueOperations;
 import org.springframework.stereotype.Component;
 
 import java.time.Duration;
-import java.util.List;
 
 @Component
 @Slf4j
@@ -25,7 +22,7 @@ public class MemberCacheAdapter {
     }
 
     public void put(String key, MemberInfoResponse value) {
-        memberCacheOperation.set(key, value, Duration.ofSeconds(24 * 60 * 60));
+        memberCacheOperation.set(key, value, Duration.ofSeconds((long)24 * 60 * 60));
     }
 
     public MemberInfoResponse get(String key) {
